myproject/employees/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import Employee, Department


class EmployeeSerializer(ModelSerializer):
    class Meta:
        model = Employee
        fields = '__all__'


class DepartmentSerializer(ModelSerializer):
    class Meta:
        model = Department
        fields = '__all__'


